requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
